move all i18n config to the same file path
