+++ src/components/locale/LocaleSwitcher.tsx
@@ -1,6 +1,6 @@
 import { useLocale, useTranslations } from "next-intl";
 import LocaleSwitcherSelect from "./LocaleSwitcherSelect";
-import { locales } from "@/i18n/config";
+import { locales } from "@/server/i18n/config";
 
 export default function LocaleSwitcher() {
   const t = useTranslations("LocaleSwitcher");
