+++ src/server/i18n/service.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { cookies } from "next/headers";
-import { Locale, defaultLocale } from "@/i18n/config";
+import { Locale, defaultLocale } from "@/server/i18n/config";
 
 // In this example the locale is read from a cookie. You could alternatively
 // also read it from a database, backend service, or any other source.
@@ -9,7 +9,7 @@ const COOKIE_NAME = "NEXT_LOCALE";
 
 export async function getUserLocale() {
   const c = await cookies();
-  return c.get(COOKIE_NAME)?.value || defaultLocale;
+  return c.get(COOKIE_NAME)?.value ?? defaultLocale;
 }
 
 export async function setUserLocale(locale: Locale) {
